+++ CHANGES.md
@@ -17,6 +17,7 @@
 * Added support for Tile Map Service (TMS) raster overlays.
 * Added support for Web Map Tile Service (WMTS) raster overlays.
 * Added raster overlay options: `maximumScreenSpaceError`, `maximumTextureSize`, `maximumSimultaneousTileLoads`, `subTileCacheBytes`.
+* Changed to use UrlAssetAccessor from vsgCs, which does a better job of managing libcurl handles than the existing HttpAssetAccessor.
 
 ### v0.17.0 - 2024-02-01